Fixes to allow compilation in clp-ffi-java.
