+++ components/core/CMakeLists.txt
@@ -691,6 +691,7 @@ set(SOURCE_FILES_unitTest
         src/ffi/search/CompositeWildcardToken.hpp
         src/ffi/search/ExactVariableToken.cpp
         src/ffi/search/ExactVariableToken.hpp
+        src/ffi/search/fmtlib_specializations.hpp
         src/ffi/search/query_methods.cpp
         src/ffi/search/query_methods.hpp
         src/ffi/search/QueryMethodFailed.hpp
